# docs/swagger/openapi.json
{
    "info": {
        "title": "Raccoon",
        "version": "1.0.0",
        "description": "Raccoon API"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/dicom-web/studies/{studyUID}": {
            "get": {
                "tags": [
                    "WADO-RS"
                ],
                "description": "Retrieve Study's instances",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/studyUID"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/MultipartRelatedDICOM"
                    }
                }
            }
        },
        "/dicom-web/studies/{studyUID}/series/{seriesUID}": {
            "get": {
                "tags": [
                    "WADO-RS"
                ],
                "description": "Retrieve Study's series' instances",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/studyUID"
                    },
                    {
                        "$ref": "#/components/parameters/seriesUID"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/MultipartRelatedDICOM"
                    }
                }
            }
        },
        "/dicom-web/studies/{studyUID}/series/{seriesUID}/instances/{instanceUID}": {
            "get": {
                "tags": [
                    "WADO-RS"
                ],
                "description": "Retrieve Study's instances",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/studyUID"
                    },
                    {
                        "$ref": "#/components/parameters/seriesUID"
                    },
                    {
                        "$ref": "#/components/parameters/instanceUID"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/MultipartRelatedDICOM"
                    }
                }
            }
        },
...
    },
    "components": {
        "responses": {
            "MultipartRelatedDICOM": {
                "description": "The response payload for WADO-RS",
                "content": {
                    "multipart/related; type=\"application/dicom\"": {
                        "schema": {
                            "type": "string",
                            "format": "binary",
                            "example": "--boundary-raccoon\r\nContent-Type: application/dicom\r\nContent-Length: {length}\r\nContent-Location: http://example.com/dicom-web/studies/{studyUID}/series/{seriesUID}/instances/{instanceUID}\r\n\r\n{data}\r\n--boundary-raccoon--"
                        }
                    }
                }
            },
...
        },
        "parameters": {
...
            "studyUID": {
                "name": "studyUID",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "seriesUID": {
                "name": "seriesUID",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "instanceUID": {
                "name": "instanceUID",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
        },
...
    },
...
}